+++ resources/views/back/_fields/document.blade.php
@@ -24,5 +24,15 @@
 @endforeach
 
 <label for="">Voeg document toe:</label>
-<input type="file" name="files[{{ $key }}][new][]" multiple style="opacity:1; position:static;"/>
-<input type="hidden" id="removeFile" name="files[{{ $key }}][delete][]" multiple/>
+<input type="file" name="files[{{ $key }}][new][]" {{ $field->multiple ? 'multiple' : '' }} style="opacity:1; position:static;"/>
+<input type="hidden" id="removeFile" name="files[{{ $key }}][delete][]" {{ $field->multiple ? 'multiple' : '' }} value=""/>
+
+@push('custom-scripts')
+<script>
+    function removeFile(id)
+    {
+        document.getElementById('removeFile').value = id;
+        document.getElementById('asset-'+id).remove();
+    }
+</script>
+@endpush