+++ resources/views/back/managers/edit.blade.php
@@ -6,14 +6,21 @@
 @component('chief::back._layouts._partials.header')
     @slot('title', $manager->details()->title)
     @slot('subtitle')
-        <a class="center-y" href="{{ $manager->route('index') }}"><span class="icon icon-arrow-left"></span> Terug naar alle {{ $manager->details()->plural }}</a>
+        <div class="inline-block">
+            <a class="center-y" href="{{ $manager->route('index') }}"><span class="icon icon-arrow-left"></span> Terug naar alle {{ $manager->details()->plural }}</a>
+        </div>
     @endslot
 
     <div class="inline-group-s">
-        {!! contract($manager, \Thinktomorrow\Chief\Management\ManagerThatPublishes::class) ? $manager->publicationStatusAsLabel() : '' !!}
 
-        <button data-submit-form="updateForm" type="button" class="btn btn-primary">Wijzigingen opslaan</button>
-        @include('chief::back.managers._partials.context-menu')
+        @if($manager->isAssistedBy('publish'))
+            {!! $manager->assistant('publish')->publicationStatusAsLabel() !!}
+        @endif
+
+        @if($manager->can('update'))
+            <button data-submit-form="updateForm" type="button" class="btn btn-primary">Wijzigingen opslaan</button>
+            @include('chief::back.managers._partials.context-menu')
+        @endif
     </div>
 
 @endcomponent
@@ -23,23 +30,23 @@
 
         <!-- needs to be before form to be detected by context-menu. Don't know why :s -->
         @include('chief::back.managers._partials.delete-modal')
-
         <form id="updateForm" method="POST" action="{{ $manager->route('update') }}" enctype="multipart/form-data" role="form">
             {{ csrf_field() }}
-            
+
             <input type="hidden" name="_method" value="PUT">
 
-            @include('chief::back.managers._partials._form', [
-                'fieldArrangement' => $manager->fieldArrangement('edit')
-             ])
+                @include('chief::back.managers._partials._form', [
+                    'fieldArrangement' => $manager->fieldArrangement('edit')
+                ])
 
-            <div class="stack text-right">
-                <button type="submit" class="btn btn-primary"> Wijzigingen opslaan</button>
-            </div>
+            @if($manager->can('update'))
+                <div class="stack text-right">
+                    <button type="submit" class="btn btn-primary">Wijzigingen opslaan</button>
+                </div>
+            @endif
 
         </form>
     </div>
-
 @stop
 
 @push('custom-styles')