+++ resources/views/back/managers/index.blade.php
@@ -24,7 +24,7 @@
         </div>
     @endif
 
-    @if(!$modelManager::filters()->isEmpty() || $modelManager::sections()->has('sidebar'))
+    @if((($managers->isEmpty() && !collect(request()->query())->isEmpty()) || (!$managers->isEmpty())) && (!$modelManager::filters()->isEmpty() || $modelManager::sections()->has('sidebar')))
         <div class="row gutter">
             <div class="column-9">
                 <div class="row gutter-s">